+++ src/pages/about.jsx
@@ -26,7 +26,7 @@ export default function About() {
           <span className='underline'>Disclaimer:</span> Any location data used is only sent to the weather API when search by location is chosen, on the map it is only kept on the client side. Analytics are also sent to a server.
           </p>
           <p className="md:absolute text-base underline md:bottom-12 md:right-1.5 md:left-auto my-5 md:my-0">
-            Version 1.0.7
+            Version 1.0.8
           </p>
       </main>
       <Footer />